TCP Improvements!
Got retransmission with TCP sort of working

+++ app/src/main/java/com/ezturner/speakersync/network/packets/tcp/TCPSeekPacket.java
@@ -31,7 +31,6 @@ public class TCPSeekPacket {
                 e.printStackTrace();
             }
         }
-
     }
 
     private void receive(InputStream stream){
@@ -45,7 +44,7 @@ public class TCPSeekPacket {
             }
         }
 
-        mSeekTime = ByteBuffer.wrap(data).getInt();
+        mSeekTime = ByteBuffer.wrap(data).getLong();
 
     }
